ref(build): dedup folder/cache bookkeeping and name memo-key separators
Three clean-ups of duplication and magic constants the perf batch grew:

- PairwiseLiteralFoldingTrait: LiteralStringFolder and LiteralArithmeticFolder
  carried verbatim copies of the n-ary pairwise-fold loop (allEqual /
  compareAll) and an identical negate(). The string folder's allEqual is just
  the === special case of compareAll, so both now share foldPairwise(callable)
  + negate().
- DeferredFlushTrait: CompiledCodeCache and PhpScanIndexCache had the same
  dirty-flag + lazy register_shutdown_function bookkeeping copy-pasted (under
  differently-named flags). Extracted to one trait; each cache keeps its own
  save() body.
- DependenciesForNamespace::memoKey: the \0 / \x01 separator bytes are now
  named constants (MEMO_ITEM_SEPARATOR / MEMO_PART_SEPARATOR) with the
  collision rationale documented, instead of bare magic bytes.

No behaviour change. Full unit/integration/core suites stay green.
